pyproject.toml
[project]
name = "employee-catalog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.1",
    "mimesis>=19.1.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "tabulate>=0.10.0",
]

[project.scripts]
edb = "modules.pgdriver.operations:edb"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["modules"]